+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## dev
+- fix `SignalsAPI.daily_model_performances`
+
 ## [2.13.0] - 2022-12-16 - "API deprecations"
 - `stake_get` - updated to migrate away from deprecated API endpoint
 - remove deprecated `SignalsAPI.daily_user_performances`